requirements.txt
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
flake8==3.9.2
pycodestyle==2.7.0
pyflakes==2.3.1
python-environ==0.4.54
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
isort==5.8.0
django-debug-toolbar==3.2.1
pytest-django==4.3.0
drf-spectacular==0.16.0
python-dotenv==0.17.1
psycopg2-binary==2.8.6
Pillow==8.2.0
django-admin-sortable2==1.0